# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is a PLAccounting API Server.",
        "title": "PLAccounting API",
        "contact": {},
        "version": "v1"
    },
    "paths": {
        "/book": {
            "get": {
                "description": "Get All Books",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Book"
                ],
                "summary": "Get All Books",
                "responses": {
                    "200": {
                        "description": "Get All Books",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "Request is failed",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            },
            "post": {
                "description": "Create Book",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Book"
                ],
                "summary": "Create Book",
                "parameters": [
                    {
                        "description": "Create Book",
                        "name": "book",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/endpoint.CreateBookRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Created Book",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "Request is failed",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/book/{bid}": {
            "get": {
                "description": "Get Book",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Book"
                ],
                "summary": "Get Book",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Book ID",
                        "name": "bid",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Get Book",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "Request is failed",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            },
            "delete": {
                "description": "Delete Book",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Book"
                ],
                "summary": "Delete Book",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Book ID",
                        "name": "bid",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Delete Book",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "Request is failed",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            },
            "patch": {
                "description": "Update Book",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Book"
                ],
                "summary": "Update Book",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Book ID",
                        "name": "bid",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Update Book",
                        "name": "book",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/endpoint.UpdateBookRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Update Book",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "Request is failed",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/book/{bid}/accountTitle": {
            "get": {
                "description": "Get All Account Titles",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Account Title"
                ],
                "summary": "Get All Account Titles",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Book ID",
                        "name": "bid",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Get All Account Titles",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "Request is failed",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            },
            "post": {
                "description": "Create Account Title",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Account Title"
                ],
                "summary": "Create Account Title",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Book ID",
                        "name": "bid",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Create Account Title",
                        "name": "accountTitle",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/endpoint.CreateAccountTitleRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Create Account Title",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "Request is failed",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/book/{bid}/accountTitle/{tid}": {
            "get": {
                "description": "Get Account Title",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Account Title"
                ],
                "summary": "Get Account Title",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Book ID",
                        "name": "bid",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Account Title ID",
                        "name": "tid",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Get Account Title",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "Request is failed",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            },
            "delete": {
                "description": "Delete Account Title",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Account Title"
                ],
                "summary": "Delete Account Title",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Book ID",
                        "name": "bid",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Account Title ID",
                        "name": "tid",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Delete Account Title",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "Request is failed",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            },
            "patch": {
                "description": "Update Account Title",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Account Title"
                ],
                "summary": "Update Account Title",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Book ID",
                        "name": "bid",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Account Title ID",
                        "name": "tid",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Update Account Title",
                        "name": "accountTitle",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/endpoint.UpdateAccountTitleRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Update Account Title",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "Request is failed",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "endpoint.CreateAccountTitleRequest": {
            "type": "object",
            "required": [
                "name"
            ],
            "properties": {
                "amount": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                },
                "type": {
                    "type": "integer"
                }
            }
        },
        "endpoint.CreateBookRequest": {
            "type": "object",
            "required": [
                "name",
                "year"
            ],
            "properties": {
                "name": {
                    "type": "string"
                },
                "year": {
                    "type": "integer"
                }
            }
        },
...
        "endpoint.UpdateAccountTitleRequest": {
            "type": "object",
            "properties": {
                "amount": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                },
                "type": {
                    "type": "integer"
                }
            }
        },
        "endpoint.UpdateBookRequest": {
            "type": "object",
            "properties": {
                "name": {
                    "type": "string"
                },
                "year": {
                    "type": "integer"
                }
            }
        },
...
    }
}